refactor: Like 컴포넌트를 공통으로 추출
- 개인이 만든 음악, 앨범 페이지 내에서도 쓰일 것이라 판단
- 이를 따로 추출하고 이름변경
