+++ app/src/main/res/layout/fragment_like_album.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.home.profile.like.AlbumLikeFragment">
+    tools:context=".presentation.home.profile.owner.AlbumLikeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_like_album"
