+++ app/src/main/res/layout/fragment_like_music.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.home.profile.like.MusicLikeFragment">
+    tools:context=".presentation.home.profile.common.MusicListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_like_music"
